refactor: re-organize generateEmoji functon
