+++ lib/config/statements.js
@@ -1,0 +1,19 @@
+const statements = [
+    'You\'ve got skills on',
+    'Wow! You really like',
+    'Good job on',
+    'Keep going on',
+    'Awesome! You must like',
+    'Excellent! Keep up the good work on'
+];
+
+const mysteryStatements = [
+    'Wow, you like the mystery language',
+    'Wow, you are a mystery man!',
+    'You must be like forking someone`s repo!'
+];
+
+module.exports = {
+    statements,
+    mysteryStatements
+};
