+++ lib/controller.js
@@ -3,9 +3,14 @@ const _ = require('lodash');
 const axios = require('axios');
 
 const colorList = require('./public/json/colors.json');
+
+// config related variables
 const error_messages = require('./config/error_messages');
-const { getEmoji, statements, mysteryStatements } = require('../lib/public/js/statements');
+const { statements, mysteryStatements } = require('./config/statements');
 const { GITHUB_API_URL } = require('./config/api');
+
+const { generateEmoji } = require('./public/js/emoji_generator');
+
 
 const client_id = process.env.CLIENT_ID || false;
 const client_secret = process.env.CLIENT_SECRET || false;
@@ -84,7 +89,7 @@ exports.index = (req, res) => {
                     languages: {},
                     msg,
                     repos,
-                    showEmoji: getEmoji(numberOfRepos),
+                    showEmoji: generateEmoji(numberOfRepos),
                     statement,
                     currentYear,
                     title
@@ -132,7 +137,7 @@ exports.index = (req, res) => {
                         avatar,
                         languages: JSON.stringify(languages),
                         colors: JSON.stringify(getColors(languages)),
-                        showEmoji: getEmoji(numberOfRepos),
+                        showEmoji: generateEmoji(numberOfRepos),
                         msg,
                         repos,
                         statement,
