+++ lib/public/css/screen/base.css
@@ -8,9 +8,7 @@ body {
   width: 100%;
   height: 100%;
   background-image: radial-gradient(circle, #0f229e 0, #202a6b 120%);
-  font-family: "proxima-nova", -apple-system, BlinkMacSystemFont, "Segoe UI",
- "Roboto", "Oxygen", "Ubuntu", "Cantarell", "Open Sans", "Helvetica Neue",
- "Arial", sans-serif; }
+  font-family: "proxima-nova", -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen", "Ubuntu", "Cantarell", "Open Sans", "Helvetica Neue", "Arial", sans-serif; }
 
 #social {
   display: inline-flex;
@@ -136,14 +134,7 @@ button.funbtn {
 
 button.funbtn:active {
   border: 1px solid orange; }
-button.funbtn.active::before {
-  content: '\f205';
-  font-family: fontAwesome;
-}
-button.funbtn.inactive::before {
-  content: '\f204';
-  font-family: fontAwesome;
-}
+
 .active {
   background-color: #37e559;
   border: 1px solid #23993a; }
