+++ lib/public/js/emoji_generator.js
@@ -1,0 +1,22 @@
+
+const generateEmoji = (repoLength) => {
+    if (repoLength >= 100) {
+        return '💯 👍 😎 👏';
+    } else if (repoLength >= 75) {
+        return '👍 😎 👏';
+    } else if (repoLength >= 50) {
+        return '👍 😎';
+    } else if (repoLength >= 20) {
+        return '👍';
+    } else if (repoLength > 0) {
+        return '🙂';
+    } else if (repoLength === 0) {
+        return '😪';
+    }
+
+    return '';
+};
+
+module.exports = {
+    generateEmoji
+};
